student can add review for a course, courses and their reviews can be searched for, fixed some mapping
