+++ ClassRegistration/ClassRegistration.App/appsettings.json
@@ -7,4 +7,8 @@
     }
   },
   "AllowedHosts": "*"
+
+
+  
+
 }
