+++ ClassRegistration/ClassRegistration.DataAccess/Entities/College.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace ClassRegistration.DataAccess.Entities
+{
+    public partial class College : DataModel
+    {
+        public College()
+        {
+            Department = new HashSet<Department>();
+        }
+
+        public int CollegeId { get; set; }
+        public string CollegeName { get; set; }
+        public string Address { get; set; }
+
+        public virtual ICollection<Department> Department { get; set; }
+    }
+}
