+++ ClassRegistration/ClassRegistration.DataAccess/Entities/Course.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace ClassRegistration.DataAccess.Entities
+{
+    public partial class Course : DataModel
+    {
+        public Course()
+        {
+            Reviews = new HashSet<Reviews>();
+            Section = new HashSet<Section>();
+        }
+
+        public int CourseId { get; set; }
+        public string CourseName { get; set; }
+        public int DeptId { get; set; }
+        public int? Credits { get; set; }
+        public decimal Fees { get; set; }
+
+        public virtual Department Dept { get; set; }
+        public virtual ICollection<Reviews> Reviews { get; set; }
+        public virtual ICollection<Section> Section { get; set; }
+    }
+}
