+++ ClassRegistration/ClassRegistration.DataAccess/Entities/DataModel.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ClassRegistration.DataAccess.Entities
+{
+    public class DataModel { }
+}
