+++ ClassRegistration/ClassRegistration.DataAccess/Entity/College.cs
@@ -1,19 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
+﻿//using System;
+//using System.Collections.Generic;
 
-namespace ClassRegistration.DataAccess.Entity
-{
-    public partial class College : DataModel
-    {
-        public College()
-        {
-            Department = new HashSet<Department>();
-        }
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public partial class College : DataModel
+//    {
+//        public College()
+//        {
+//            Department = new HashSet<Department>();
+//        }
 
-        public int CollegeId { get; set; }
-        public string CollegeName { get; set; }
-        public string Address { get; set; }
+//        public int CollegeId { get; set; }
+//        public string CollegeName { get; set; }
+//        public string Address { get; set; }
 
-        public virtual ICollection<Department> Department { get; set; }
-    }
-}
+//        public virtual ICollection<Department> Department { get; set; }
+//    }
+//}
