+++ ClassRegistration/ClassRegistration.DataAccess/Entity/DataModel.cs
@@ -1,8 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Text;
 
-namespace ClassRegistration.DataAccess.Entity
-{
-    public class DataModel { }
-}
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public class DataModel { }
+//}
