+++ ClassRegistration/ClassRegistration.DataAccess/Entity/Enrollment.cs
@@ -1,15 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
+﻿//using System;
+//using System.Collections.Generic;
 
-namespace ClassRegistration.DataAccess.Entity
-{
-    public partial class Enrollment : DataModel
-    {
-        public int EnrollmentId { get; set; }
-        public int StudentId { get; set; }
-        public int SectId { get; set; }
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public partial class Enrollment : DataModel
+//    {
+//        public int EnrollmentId { get; set; }
+//        public int StudentId { get; set; }
+//        public int SectId { get; set; }
 
-        public virtual Section Sect { get; set; }
-        public virtual Student Student { get; set; }
-    }
-}
+//        public virtual Section Sect { get; set; }
+//        public virtual Student Student { get; set; }
+//    }
+//}
