+++ ClassRegistration/ClassRegistration.DataAccess/Entity/Instructor.cs
@@ -1,21 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
+﻿//using System;
+//using System.Collections.Generic;
 
-namespace ClassRegistration.DataAccess.Entity
-{
-    public partial class Instructor : DataModel
-    {
-        public Instructor()
-        {
-            Section = new HashSet<Section>();
-        }
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public partial class Instructor : DataModel
+//    {
+//        public Instructor()
+//        {
+//            Section = new HashSet<Section>();
+//        }
 
-        public int InstructorId { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public int DeptId { get; set; }
+//        public int InstructorId { get; set; }
+//        public string FirstName { get; set; }
+//        public string LastName { get; set; }
+//        public int DeptId { get; set; }
 
-        public virtual Department Dept { get; set; }
-        public virtual ICollection<Section> Section { get; set; }
-    }
-}
+//        public virtual Department Dept { get; set; }
+//        public virtual ICollection<Section> Section { get; set; }
+//    }
+//}
