+++ ClassRegistration/ClassRegistration.DataAccess/Entity/Reviews.cs
@@ -1,0 +1,25 @@
+﻿//using System;
+//using System.Collections.Generic;
+
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public class Reviews : DataModel
+//    {
+//        public Reviews()
+//        {
+//        }
+
+//        public int ReviewId { get; set; }
+//        public int Score { get; set; }
+//        public string Text { get; set; }
+//        public DateTime Date { get; set; }
+//        public int StudentId { get; set; }
+//        public int CourseId { get; set; }
+
+//        public virtual Course Course { get; set; }
+//        public virtual Student Student { get; set; }
+//    }
+//}
+
+
+
