+++ ClassRegistration/ClassRegistration.DataAccess/Entity/Section.cs
@@ -1,25 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
+﻿//using System;
+//using System.Collections.Generic;
 
-namespace ClassRegistration.DataAccess.Entity
-{
-    public partial class Section : DataModel
-    {
-        public Section()
-        {
-            Enrollment = new HashSet<Enrollment>();
-        }
+//namespace ClassRegistration.DataAccess.Entity
+//{
+//    public partial class Section : DataModel
+//    {
+//        public Section()
+//        {
+//            Enrollment = new HashSet<Enrollment>();
+//        }
 
-        public int SectId { get; set; }
-        public int CourseId { get; set; }
-        public int InstructorId { get; set; }
-        public string Term { get; set; }
-        public TimeSpan StartTime { get; set; }
-        public TimeSpan EndTime { get; set; }
+//        public int SectId { get; set; }
+//        public int CourseId { get; set; }
+//        public int InstructorId { get; set; }
+//        public string Term { get; set; }
+//        public TimeSpan StartTime { get; set; }
+//        public TimeSpan EndTime { get; set; }
 
-        public virtual Course Course { get; set; }
-        public virtual Instructor Instructor { get; set; }
-        public virtual Semester TermNavigation { get; set; }
-        public virtual ICollection<Enrollment> Enrollment { get; set; }
-    }
-}
+//        public virtual Course Course { get; set; }
+//        public virtual Instructor Instructor { get; set; }
+//        public virtual Semester TermNavigation { get; set; }
+//        public virtual ICollection<Enrollment> Enrollment { get; set; }
+//    }
+//}
