+++ ClassRegistration/ClassRegistration.DataAccess/Interfaces/IStudentRepository.cs
@@ -5,6 +5,12 @@ namespace ClassRegistration.Domain
 {
     public interface IStudentRepository
     {
+
+        /// <summary>
+        /// Search a student by ID
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         Task<StudentModel> FindById (int id);
     }
 }
