+++ ClassRegistration/ClassRegistration.DataAccess/Repository/SectionRepository.cs
@@ -1,4 +1,4 @@
-﻿using ClassRegistration.DataAccess.Entity;
+﻿using ClassRegistration.DataAccess.Entities;
 using ClassRegistration.DataAccess.Interfaces;
 using ClassRegistration.Domain.Model;
 using Microsoft.EntityFrameworkCore;
@@ -10,24 +10,54 @@ namespace ClassRegistration.DataAccess.Repository
 {
     public class SectionRepository : Repository<Section, SectionModel>, ISectionRepository
     {
-        public SectionRepository (Course_registration_dbContext context) : base (context) { }
+        public SectionRepository(Course_registration_dbContext context) : base(context) { }
 
-        // get access to course navigation properties given an instructor ID
-        public virtual async Task<IEnumerable<SectionModel>> FindByInstrId (int instructorId)
-        {
-            // TODO we want to make stuff like this use the LINQ query syntax
-            var sections = await _context.Section.Include (s => s.Course)
-                                    .Where (s => s.InstructorId == instructorId).ToListAsync ();
+        public SectionRepository() : this(null) { }
 
-            return _mapper.Map<IEnumerable<SectionModel>> (sections);
-        }
 
+        /// <summary>
+        /// All the sections available
+        /// </summary>
+        /// <returns></returns>
         // get all the sections : just to verify in postman
-        public virtual async Task<IEnumerable<SectionModel>> FindAll ()
+        public virtual async Task<IEnumerable<SectionModel>> FindAll()
         {
-            //    var class = _dbContext.Course.Where(c => c.CourseId == courseId);
-            var sections = await _context.Section.ToListAsync ();
-            return _mapper.Map<IEnumerable<SectionModel>> (sections);
+
+            var sections = await _context.Section.ToListAsync();
+
+            return _mapper.Map<IEnumerable<SectionModel>>(sections);
         }
+
+
+
+        /// <summary>
+        /// Find the courses and their reveiws by instructor ID
+        /// </summary>
+        /// <param name="instructorId"></param>
+        /// <returns></returns>
+        // get access to course navigation properties given an instructor ID
+        public virtual async Task<IEnumerable<Section>> FindByInstrId (int instructorId)
+        {
+            // we want to make stuff like this use the LINQ query syntax
+            var sections = await _context.Section
+               .Include(s => s.Course)
+                .ThenInclude(c => c.Reviews)
+               .Where(s => s.InstructorId == instructorId).ToListAsync();
+
+            return sections;
+            
+
+        }
+
     }
 }
+
+       
+    
+
+
+        
+
+
+
+
