+++ ClassRegistration/ClassRegistration.Domain/Model/CourseModel.cs
@@ -5,10 +5,11 @@ namespace ClassRegistration.Domain.Model
 {
     public class CourseModel : BaseBusinessModel
     {
-        public CourseModel ()
-        {
-            Section = new HashSet<SectionModel> ();
-        }
+        //public Course ()
+        //{
+        //    Section = new HashSet<Section> ();
+        //    Reviews = new HashSet<ReviewsModel>();
+        //}
 
         private string _courseName;
 
@@ -31,6 +32,8 @@ namespace ClassRegistration.Domain.Model
         public int? Credits { get; set; }
         public decimal Fees { get; set; }
 
-        public virtual ICollection<SectionModel> Section { get; set; }
+        public List<SectionModel> Section { get; set; } = new List<SectionModel>();
+
+        public virtual ICollection<ReviewsModel> Reviews { get; set; }
     }
 }
