+++ ClassRegistration/ClassRegistration.Domain/Model/SectionModel.cs
@@ -1,9 +1,16 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace ClassRegistration.Domain.Model
 {
     public class SectionModel : BaseBusinessModel
     {
+
+        //public SectionModel()
+        //{
+        //    Enrollment = new HashSet<Enrollment>();
+        //}
+
         public int SectId { get; set; }
         public int CourseId { get; set; }
         public int InstructorId { get; set; }
@@ -11,6 +18,10 @@ namespace ClassRegistration.Domain.Model
         public TimeSpan StartTime { get; set; }
         public TimeSpan EndTime { get; set; }
 
-        public virtual CourseModel Course { get; set; }
+        public CourseModel Course { get; set; } 
+        //public virtual InstructorModel Instructor { get; set; }
+        //public virtual SemesterModel Semester { get; set; }
+        //public virtual ICollection<Enrollment> Enrollment { get; set; }
     }
 }
+
